pyproject.toml
[tool.poetry]
name = "forex-backend"
version = "0.1.0"
description = "Backend API for forex data collection and analysis"
authors = ["Your Name <you@example.com>"]
license = "MIT"
package-mode = false
packages = [
    { include = "app" }
]



[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.136.0"
uvicorn = "^0.45.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.49"}
asyncpg = "^0.31.0"
alembic = "^1.18.4"
python-dotenv = "^1.2.2"
pydantic = "^2.13.3"
httpx = "^0.28.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"
pytest-asyncio = "^1.3.0"
ruff = "^0.11.0"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

